requirements.txt
# torch==1.7.0+cu110
# torchvision==0.8.1+cu110
pandas==1.0.1
numpy==1.18.1
Pillow==8.3.1
scikit_learn==0.24.2
